removed conversion to long.

+++ Model.py
@@ -32,7 +32,7 @@ class Wavenet:
 
     def train(self, input, real):
         output = self.net(input)
-        loss = self.loss(output.view(-1, self.in_channels), real.long().view(-1, self.in_channels))
+        loss = self.loss(output.view(-1, self.in_channels), real.view(-1, self.in_channels))
         self.optimizer.zero_grad()
         loss.backward()
         self.optimizer.step()
